test-renderers: Dynamically link the shared code, and split crates.
* The library formerly known as `test-renderers` is now the dynamically
  linked `test-renderers-dylib`. This reduces the build time (and,
  presumably, binary size, though I haven’t checked) for the 3 renderer
  test targets by avoiding linking the same code 3 times.

* I split the library into 3: `test-renderers-types` contains shared
  code, `test-renderers-cases` contains only the test cases, and
  `test-renderers-runner` contains test filtering and report generation.
  This shortens the critical path for `test-renderers-dylib` and reduces
  how much needs to be rebuilt if the cases are edited.

* Also removed some unused dependencies.
