+++ Cargo.toml
@@ -16,6 +16,10 @@ members = [
   "test-aic",
   "test-mesh",
   "test-renderers",
+  "test-renderers/cases",
+  "test-renderers/dylib",
+  "test-renderers/runner",
+  "test-renderers/types",
   "tools/xtask",
 
   # Workspaces generally expect to be able to compile everything for the same target,
@@ -304,6 +308,9 @@ package.all-is-cubes-ui.opt-level = 1
 package.test-aic.opt-level = 0
 package.test-mesh.opt-level = 0
 package.test-renderers.opt-level = 0
+package.test-renderers-cases.opt-level = 0
+package.test-renderers-runner.opt-level = 0
+package.test-renderers-types.opt-level = 0
 package.xtask.opt-level = 0
 
 [profile.release]
