+++ clippy.toml
@@ -7,7 +7,7 @@ enforced-import-renames = [
     { path = "all_is_cubes::rerun_glue", rename = "rg" },
 ]
 ignore-interior-mutability = [
-    "test_renderers::harness::UniverseFuture", # contains futures::Shared but does not use it as key
+    "test_renderers_runner::harness::UniverseFuture", # contains futures::Shared but does not use it as key
 ]
 
 # This configures the `modulo_arithmetic` lint.
