+++ test-renderers/types/src/comparison.rs
@@ -1,132 +1,31 @@
-//! Test harness for [`all_is_cubes`]’ various renderers, supporting comparing the
-//! image outputs of various renderers with each other and reference images.
-
-// Crate-specific lint settings. (General settings can be found in the workspace manifest.)
-#![allow(
-    missing_docs,
-    clippy::module_name_repetitions,
-    clippy::exhaustive_enums,
-    clippy::exhaustive_structs,
-    clippy::missing_panics_doc,
-    reason = "library for internal use only"
-)]
-#![forbid(unsafe_code)]
-
-use std::fmt;
 use std::fs;
-use std::io;
-use std::path::Path;
-use std::path::PathBuf;
+use std::io::{self};
+use std::path::{Path, PathBuf};
 use std::time::Instant;
 
-use clap::builder::PossibleValue;
-use image::RgbaImage;
 use rendiff::{Histogram, Threshold};
 
-use all_is_cubes::character::Character;
 use all_is_cubes::euclid::size2;
-use all_is_cubes::space::Space;
-use all_is_cubes::universe::{Handle, Universe};
 use all_is_cubes::util::{ConciseDebug, Refmt as _, StatusText};
-use all_is_cubes_content::free_editing_starter_inventory;
 use all_is_cubes_render::Rendering;
 use all_is_cubes_render::camera::ImageSize;
 
-mod harness;
-pub use harness::*;
-mod image_files;
-pub use image_files::*;
-mod render;
-pub use render::*;
-mod report;
-pub use report::*;
-pub mod test_cases;
+use crate::{ImageId, NotFound, RendererId, Version, image_path, load_and_copy_expected_image};
 
-/// A suite is a group of tests which get their own image directories, report files, and
-/// test targets.
-#[derive(
-    Copy, Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, serde::Serialize, serde::Deserialize,
-)]
-#[expect(clippy::exhaustive_enums)]
-pub enum SuiteId {
-    Renderers,
-    Ui,
-}
-impl SuiteId {
-    fn as_str(self) -> &'static str {
-        match self {
-            SuiteId::Renderers => "renderers",
-            SuiteId::Ui => "ui",
-        }
-    }
-}
-impl clap::ValueEnum for SuiteId {
-    fn value_variants<'a>() -> &'a [Self] {
-        &[SuiteId::Renderers, SuiteId::Ui]
-    }
-
-    fn to_possible_value(&self) -> Option<PossibleValue> {
-        Some(PossibleValue::new(self.as_str()))
-    }
-}
-impl fmt::Display for SuiteId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        f.write_str(self.as_str())
-    }
-}
-
-#[derive(
-    Clone, Debug, Eq, Hash, Ord, PartialEq, PartialOrd, serde::Serialize, serde::Deserialize,
-)]
-#[expect(clippy::exhaustive_structs)]
-pub struct TestId {
-    pub suite: SuiteId,
-
-    /// Name of the test function and parameters, used as part of the image file name.
-    pub test: String,
-}
-
-impl fmt::Display for TestId {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let Self { suite, test } = self;
-        write!(f, "{suite}/{test}")
-    }
-}
-
-/// Given a [`Space`], create the [`Character`] looking at it, with the default name.
-pub fn finish_universe_from_space(
-    universe: &mut Universe,
-    space: Space,
-) -> (Handle<Space>, Handle<Character>) {
-    // TODO: "character".into() shouldn't be sprinkled around various universe construction.
-    let space_handle = universe.insert("space".into(), space).unwrap();
-
-    // If the universe is dumped, include tools and jetpack
-    let mut spawn = space_handle.read(universe.read_ticket()).unwrap().spawn().clone();
-    spawn.set_inventory(free_editing_starter_inventory(true));
-
-    let character_handle = universe
-        .insert(
-            "character".into(),
-            Character::spawn(&spawn, space_handle.clone()),
-        )
-        .unwrap();
-
-    (space_handle, character_handle)
-}
+// -------------------------------------------------------------------------------------------------
 
 /// Result of calling [`compare_rendered_image`]
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
 pub struct ComparisonRecord {
-    expected_image: ComparisonImage,
-    actual_image: ComparisonImage,
-    diff_image: Option<ComparisonImage>,
-    diff_histogram: Vec<usize>, // length 256; is a Vec for serializability
-    outcome: ComparisonOutcome,
+    pub expected_image: ComparisonImage,
+    pub actual_image: ComparisonImage,
+    pub diff_image: Option<ComparisonImage>,
+    pub diff_histogram: Vec<usize>, // length 256; is a Vec for serializability
+    pub outcome: ComparisonOutcome,
 
     // Info from the renderer.
     // Not part of the comparison, but something we want to include in the report.
-    render_info: String,
+    pub render_info: String,
 }
 
 /// Information about an image file to be displayed in the report.
@@ -134,9 +33,9 @@ pub struct ComparisonRecord {
 pub struct ComparisonImage {
     /// Name of the image file (only; not an absolute path).
     /// The directory must be known by context.
-    file_name: String,
-    width: u32,
-    height: u32,
+    pub file_name: String,
+    pub width: u32,
+    pub height: u32,
 }
 
 #[derive(Clone, Debug, Eq, PartialEq, serde::Serialize, serde::Deserialize)]
@@ -156,8 +55,8 @@ pub enum ComparisonOutcome {
 }
 
 impl ComparisonOutcome {
-    /// Check if the outcome is flawed, that is, a comparison failure occurred but should be disregarded.|
-    /// This method exists to centralize a couple of cases where we make this check
+    /// Check if the outcome is flawed, that is, a comparison failure occurred but should be disregarded.
+    /// This method exists to centralize a couple of cases where we make this check.
     pub fn is_flawed(&self) -> bool {
         match self {
             ComparisonOutcome::Flawed(_) => true,
@@ -175,7 +74,7 @@ impl ComparisonRecord {
         expected_image: ComparisonImage,
         actual_image: ComparisonImage,
         diff_image: Option<ComparisonImage>,
-        diff_histogram: Histogram,
+        diff_histogram: rendiff::Histogram,
         outcome: ComparisonOutcome,
         render_info: String,
     ) -> Self {
@@ -189,7 +88,7 @@ impl ComparisonRecord {
         }
     }
 
-    fn describe_failure(&self) -> Option<String> {
+    pub fn describe_failure(&self) -> Option<String> {
         match self.outcome {
             ComparisonOutcome::Equal | ComparisonOutcome::Flawed(_) => None,
             ComparisonOutcome::Different { amount } => {
@@ -275,7 +174,7 @@ pub fn compare_rendered_image(
 
     // Save diff image to disk
     let diff_c_image: Option<ComparisonImage> = if let Some(image) = diff_result.diff_image() {
-        RgbaImage::from_raw(
+        image::RgbaImage::from_raw(
             image.width() as u32,
             image.height() as u32,
             image.buf().to_vec().into_flattened(),
@@ -338,46 +237,4 @@ pub fn compare_rendered_image(
     );
 
     record
-}
-
-pub fn initialize_logging(args: &HarnessArgs) {
-    struct RenderHarnessLogger {
-        level_filter: log::LevelFilter,
-    }
-    impl log::Log for RenderHarnessLogger {
-        fn enabled(&self, metadata: &log::Metadata<'_>) -> bool {
-            let t = metadata.target();
-            metadata.level() <= self.level_filter
-                && (!t.starts_with("wgpu") || t == "wgpu_render")
-                && !t.starts_with("naga")
-        }
-        fn log(&self, record: &log::Record<'_>) {
-            use io::Write as _;
-
-            if self.enabled(record.metadata()) {
-                let test_id: Option<TestId> = TEST_ID.try_with(|id| id.clone()).ok();
-                let mut lock = io::stderr().lock();
-                _ = time::OffsetDateTime::now_utc().to_offset(time::UtcOffset::UTC).format_into(
-                    &mut lock,
-                    time::macros::format_description!("[hour]:[minute]:[second]"),
-                );
-                _ = write!(lock, " [{level}] (", level = record.level());
-                _ = match test_id {
-                    Some(id) => write!(lock, "{id}"),
-                    None => write!(lock, "?"),
-                };
-                _ = writeln!(lock, ") {}", record.args());
-            }
-        }
-        fn flush(&self) {}
-    }
-
-    let level_filter = if args.verbose {
-        log::LevelFilter::Trace
-    } else {
-        log::LevelFilter::Warn
-    };
-
-    log::set_logger(Box::leak(Box::new(RenderHarnessLogger { level_filter }))).unwrap();
-    log::set_max_level(level_filter);
 }
